feat(cal-ui-p4): move calendar nav components (today + view switcher) into package
Move the presentational CalendarNavigation and CalendarViewSwitcher into
@microboxlabs/miot-calendar-ui and export from src/index.ts. Both are
props-driven (label/handler, active view + onViewChange) so the move is
mechanical; CalendarViewSwitcher's ViewMode import repoints to the package
service types. App keeps thin re-export shims at the old paths so
planning-header stays green.
