+++ turbo-repo/apps/app/src/features/calendar/components/planning/calendar-navigation.tsx
@@ -1,19 +1,4 @@
-"use client";
-
-import { Button } from "flowbite-react";
-
-interface CalendarNavigationProps {
-  onToday: () => void;
-  todayLabel: string;
-}
-
-export function CalendarNavigation({
-  onToday,
-  todayLabel,
-}: Readonly<CalendarNavigationProps>) {
-  return (
-    <Button color="alternative" size="sm" onClick={onToday}>
-      {todayLabel}
-    </Button>
-  );
-}
+// The CalendarNavigation "today" button now lives in
+// @microboxlabs/miot-calendar-ui. This thin shim keeps the old import path
+// resolving so existing importers (planning-header) stay green.
+export { CalendarNavigation } from "@microboxlabs/miot-calendar-ui";
