+++ turbo-repo/packages/miot-calendar-ui/src/components/planning/calendar-navigation.tsx
@@ -1,0 +1,19 @@
+"use client";
+
+import { Button } from "flowbite-react";
+
+interface CalendarNavigationProps {
+  onToday: () => void;
+  todayLabel: string;
+}
+
+export function CalendarNavigation({
+  onToday,
+  todayLabel,
+}: Readonly<CalendarNavigationProps>) {
+  return (
+    <Button color="alternative" size="sm" onClick={onToday}>
+      {todayLabel}
+    </Button>
+  );
+}
